airflow-docker/requirements.txt
alembic==1.7.5
amqp==5.0.7
anyio==3.4.0
apache-airflow==2.3.1 # Changed from 2.2.5 to 2.3.1
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Automat==20.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
cachelib==0.6.0
cattrs==1.5.0
celery==5.2.1
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.0
constantly==15.1.0
croniter==1.0.15
# cryptography==3.4.8 # Commented out
cssselect==1.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
filelock==3.6.0
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==1.0.0
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
incremental==21.3.0
inflection==0.5.1
iso8601==1.0.2
isodate==0.6.0
itemadapter==0.5.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==3.0.3
jmespath==1.0.0
jsonschema==3.2.0
kombu==5.2.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
numpy==1.21.6 # Changed from 1.22.3 to 1.21.6
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.3.5 # Changed from 1.4.2 to 1.3.5
parsel==1.6.0
pendulum==2.1.2
playwright==1.20.1
prison==0.2.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
Protego==0.2.1
proxyscrape==0.3.0
psutil==5.8.0
pyairtable==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyee==8.1.0
Pygments==2.10.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-nvd3==0.15.0
python-slugify==5.0 # Changed from 4.0.1 to 5.0
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
queuelib==1.6.2
requests==2.26.0
requests-file==1.5.1
rfc3986==1.5.0
rich==12.4.1 # Changed from 10.16.0 to 12.4.1
scraperapi-sdk==0.2.2
Scrapy==2.6.1
scrapy-playwright==0.0.14
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1
scrapy-zyte-smartproxy==2.1.0
service-identity==21.1.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2
# SQLAlchemy==1.3.24 # Commented out
# SQLAlchemy-JSONField==1.0.0 # Commented out
# SQLAlchemy-Utils==0.37.9 # Commented out
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tldextract==3.2.0
tornado==6.1
Twisted==22.2.0
typing-extensions==4.1.1
ua-parser==0.10.0
unicodecsv==0.14.1
Unidecode==1.3.4
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.1
Werkzeug==1.0.1
wrapt==1.14.0
WTForms==2.3.3
zipp==3.6.0
zope.interface==5.4.0
